+++ src/layers/utils.js
@@ -2,7 +2,7 @@ import { Matrix4 } from 'math.gl';
 import {
   MAX_COLOR_INTENSITY,
   DEFAULT_COLOR_OFF,
-  MAX_SLIDERS_AND_CHANNELS,
+  MAX_CHANNELS,
   DTYPE_VALUES
 } from '../constants';
 
@@ -30,49 +30,48 @@ export function getDtypeValues(dtype) {
   return values;
 }
 
-export function padColorsAndSliders({
-  sliderValues,
-  colorValues,
-  channelIsOn,
+export function padColorsAndWindows({
+  contrastLimits,
+  colors,
+  channelsVisible,
   domain,
   dtype
 }) {
-  const lengths = [sliderValues.length, colorValues.length];
+  const lengths = [contrastLimits.length, colors.length];
   if (lengths.every(l => l !== lengths[0])) {
     throw Error('Inconsistent number of slider values and colors provided');
   }
 
-  const colors = colorValues.map((color, i) =>
-    channelIsOn[i] ? color.map(c => c / MAX_COLOR_INTENSITY) : DEFAULT_COLOR_OFF
+  const newColors = colors.map((color, i) =>
+    channelsVisible[i]
+      ? color.map(c => c / MAX_COLOR_INTENSITY)
+      : DEFAULT_COLOR_OFF
   );
   const maxSliderValue = (domain && domain[1]) || getDtypeValues(dtype).max;
-  const sliders = sliderValues.map((slider, i) =>
-    channelIsOn[i] ? slider : [maxSliderValue, maxSliderValue]
+  const newContrastLimits = contrastLimits.map((slider, i) =>
+    channelsVisible[i] ? slider : [maxSliderValue, maxSliderValue]
   );
-  // Need to pad sliders and colors with default values (required by shader)
-  const padSize = MAX_SLIDERS_AND_CHANNELS - colors.length;
+  // Need to pad contrastLimits and colors with default values (required by shader)
+  const padSize = MAX_CHANNELS - newColors.length;
   if (padSize < 0) {
     throw Error(`${lengths} channels passed in, but only 6 are allowed.`);
   }
 
-  const paddedColorValues = padWithDefault(colors, DEFAULT_COLOR_OFF, padSize);
-  const paddedSliderValues = padWithDefault(
-    sliders,
+  const paddedColors = padWithDefault(newColors, DEFAULT_COLOR_OFF, padSize);
+  const paddedContrastLimits = padWithDefault(
+    newContrastLimits,
     [maxSliderValue, maxSliderValue],
     padSize
   );
-  const paddedColorsAndSliders = {
-    paddedSliderValues: paddedSliderValues.reduce(
+  const paddedColorsAndWindows = {
+    paddedContrastLimits: paddedContrastLimits.reduce(
       (acc, val) => acc.concat(val),
       []
     ), // flatten for use on shaders
-    paddedColorValues: paddedColorValues.reduce(
-      (acc, val) => acc.concat(val),
-      []
-    )
+    paddedColors: paddedColors.reduce((acc, val) => acc.concat(val), [])
   };
 
-  return paddedColorsAndSliders;
+  return paddedColorsAndWindows;
 }
 
 export function onPointer(layer) {