+++ tests/layers_views/layers/VolumeLayer.spec.js
@@ -17,15 +17,15 @@ test('VolumeLayer', t => {
     Layer: VolumeLayer,
     assert: t.ok,
     sampleProps: {
-      sliderValues: [
+      contrastLimits: [
         [0, 10],
         [0, 10]
       ],
-      colorValues: [
+      colors: [
         [0, 1, 1],
         [0, 1, 1]
       ],
-      channelIsOn: [true, false],
+      channelsVisible: [true, false],
       loader: [
         {
           getRaster: async () => ({
@@ -48,7 +48,7 @@ test('VolumeLayer', t => {
           dtype: 'Uint32'
         }
       ],
-      loaderSelection: [{}, {}]
+      selections: [{}, {}]
     },
     onBeforeUpdate: ({ testCase }) => t.comment(testCase.title)
   });
@@ -78,15 +78,15 @@ test('VolumeLayer', t => {
   const testCases = [
     {
       props: {
-        sliderValues: [
+        contrastLimits: [
           [0, 10],
           [0, 10]
         ],
-        colorValues: [
+        colors: [
           [0, 1, 1],
           [0, 1, 1]
         ],
-        channelIsOn: [true, false],
+        channelsVisible: [true, false],
         loader: [
           {
             getRaster: async () => {
@@ -109,7 +109,7 @@ test('VolumeLayer', t => {
             dtype: 'Uint32'
           }
         ],
-        loaderSelection: []
+        selections: []
       },
       onAfterUpdate: () => {
         t.ok(
@@ -120,7 +120,7 @@ test('VolumeLayer', t => {
     },
     {
       updateProps: {
-        loaderSelection: [1, 2]
+        selections: [1, 2]
       },
       onAfterUpdate: () =>
         t.ok(state.data.length === 8, 'Updated loader selection requests data.')