+++ tests/layers_views/layers/XRLayer.spec.js
@@ -18,9 +18,9 @@ test('XRLayer', t => {
     assert: t.ok,
     sampleProps: {
       bounds: [0, 0, 2, 2],
-      sliderValues: [[0, 10]],
-      colorValues: [[0, 1, 1]],
-      channelIsOn: [true],
+      contrastLimits: [[0, 10]],
+      colors: [[0, 1, 1]],
+      channelsVisible: [true],
       tileSize: 2,
       channelData: {
         data: [new Uint32Array([0, 2, 1, 2])],